+++ sdk/core/azure-core/inc/azure/core/http/policy.hpp
@@ -8,6 +8,7 @@
 
 #pragma once
 
+#include "azure/core/case_insensitive_map.hpp"
 #include "azure/core/context.hpp"
 #include "azure/core/credentials.hpp"
 #include "azure/core/http/http.hpp"
@@ -388,7 +389,7 @@ namespace Azure { namespace Core { namespace Http {
      */
     struct ValuePolicyOptions
     {
-      std::map<std::string, std::string> HeaderValues;
+      CaseInsensitiveMap<std::string> HeaderValues;
       std::map<std::string, std::string> QueryValues;
     };